Create: add --reserve-space support to reserve space at the start of each device
This only works for 1.x metadata.
The idea is to reserve some space at the start of each device for
storing a boot loader.
As a boot loader will need sector 0, metadata 1.1 doesn't make sense.

The number given to reserve-space will be in kilobytes, though a
suffix of M can give Megabytes (or G gigabytes, but 2 is the limit).

When a spare is added to an array, we preserve the amount of reserved
space.

It is not yet possible to hot-add an internal bitmap to a 1.1 or 1.2
array with reserved space.

This option is not yet documented.

In future, we might copy the content of the reserved space from one
drive to any new spare that is added.

Signed-off-by: NeilBrown <neilb@suse.de>